refactor(design-system): move Field out into apps/web/src/common
Field was the only design-system component bound to a specific form library
(@tanstack/react-form, previously the package's sole optional peer). Moving it
into apps/web/src/common — the app's own form adapter — leaves design-system
fully form-library-agnostic.

- New apps/web/src/common/Field/ (Field.tsx, Field.test.tsx, README.md,
  index.ts), imports switched to @soroush.tech/design-system/* and the app's
  own renderWithTheme (src/test/utils/wrapper).
- ContactInquire and the design-system demo's FormCard import Field from
  src/common/Field.
- design-system drops the ./Field export (source + publishConfig) and
  @tanstack/react-form entirely (peerDependencies, peerDependenciesMeta,
  devDependencies).
- Docs updated: design-system/README.md's component table, and Form/README.md's
  cross-link to Field's old location.

Verified with lint, typecheck, and 100% coverage (both packages, the exact CI
commands) plus a production build.
